restore the old drake visualizer interface as fallback
